classroom/grading/requirements.txt
rht-labs-core
rht-labs-dle
openshift-client
flake8-black
pytest
